Rename example + addressing Shreyas comment
