--- src/nidigital/examples/burst_with_programmatic_configuration/PinLevels.digilevels
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PinLevelsFile xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" schemaVersion="1.0" xmlns="http://www.ni.com/Semiconductor/PinLevels">
-  <PinLevelsSheet>
-    <DigitalPinLevelSets>
-      <DigitalPinLevelSet pin="PinGroup1">
-        <Vil>0</Vil>
-        <Vih>3.3</Vih>
-        <Vol>0.8</Vol>
-        <Voh>2</Voh>
-        <Vterm>0</Vterm>
-        <TerminationMode>HighZ</TerminationMode>
-      </DigitalPinLevelSet>
-    </DigitalPinLevelSets>
-  </PinLevelsSheet>
-</PinLevelsFile>
