--- src/nidigital/examples/burst_with_programmatic_configuration/Specifications.specs
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Specifications xmlns:f="http://www.ni.com/schemas/Semiconductor/Formula.xsd" schemaVersion="1.0" xmlns="http://www.ni.com/schemas/Semiconductor/Specifications.xsd">
-  <Section name="timing">
-    <f:Formula symbol="period">
-      <f:Definition>1 µs</f:Definition>
-    </f:Formula>
-  </Section>
-  <Description />
-</Specifications>
