+++ src/nidigital/examples/configure_time_set_and_voltage_levels/Pattern.digipat
